third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/intrinsic-auto-repeat/row-auto-repeat-auto-015-expected.txt

This is a testharness.js-based test.
[FAIL] Auto-fit repeat tracks with auto size and auto item placement
  assert_equals: expected "250px 250px 250px 250px 0px 0px 0px 0px 0px 0px" but got "none"
Harness: the test ran to completion.
